distro: catch ImportError when loading instrumentors entry points
When running in the k8s operator loading of an instrumentor may fail
because the environment in which python extensions are built does not
match the one from the running container.
There are at least two cases:
- different python version
- different kind of wheels, e.g. manylinux vs musllinux

To avoid the distro loading to fail catch ImportError here, that is
the kind of exception we see when loading shared objects or cython
extensions fails.