distro: override otlp exporters user agent headers (#363)
* distro: override otlp exporters user agent headers

For GRPC we leverage the recently added channel_options, for HTTP we
temporarily need to monkey patch the exporter default headers in the
meantime we have an sdk where we can override them.

* Teach our integrations utils about http tests

And add a test for http otlp exporter user-agent. The http server
requires to bind to a random port for each test because as it won't
unbind it until process exit.
And SO_REUSEPORT does not seem to help either.