Fix logging without a tty (#436)
This fixes logging without an iteractive session, e.g. docker without
-it or by default on k8s.

The issue there is that without explicitly adding handlers without a tty
the log are not sent to stderr.

Also we need to take care of the propagation in case the OTel Logging
handler is configured on the root logger to have also these logs
exported.